Move the commandline args to their own project, so Remoting can reference them too
