+++ src/cascadia/CommandlineArgs/PlaceholderType.cpp
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+#include "pch.h"
+#include "PlaceholderType.h"
+#include "PlaceholderType.g.cpp"
+
+namespace winrt::Microsoft::Terminal::CommandlineArgs::implementation
+{
+}
