+++ src/cascadia/CommandlineArgs/pch.cpp
@@ -1,0 +1,4 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+#include "pch.h"
