+++ src/cascadia/TerminalApp/CommandPalette.h
@@ -5,7 +5,7 @@
 
 #include "FilteredCommand.h"
 #include "CommandPalette.g.h"
-#include "AppCommandlineArgs.h"
+#include "../CommandlineArgs/AppCommandlineArgs.h"
 #include "../../cascadia/inc/cppwinrt_utils.h"
 
 // fwdecl unittest classes
